LogX fix, confusion, more tests
