+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/impl/transforms/LogX.java
@@ -40,6 +40,7 @@ public class LogX extends BaseTransformOp {
     public LogX(SameDiff sameDiff, SDVariable i_v, double base) {
         super(sameDiff, i_v, new Object[] {base});
         this.base = base;
+        this.extraArgs = new Object[] {base};
     }
 
     public LogX() {}
